deleted unused functions and fixed tests
